Move impl_serde_hex_for_binary_value to the Merkledb
