+++ exonum/src/messages/mod.rs
@@ -101,8 +101,6 @@ use crate::{
     helpers::{Height, Round, ValidatorId},
 };
 
-#[macro_use]
-mod macros;
 mod signed;
 #[cfg(test)]
 mod tests;
