+++ exonum/src/runtime/mod.rs
@@ -132,7 +132,7 @@
 
 pub use self::{
     blockchain_data::{BlockchainData, SnapshotExt},
-    dispatcher::{Dispatcher, Error as DispatcherError, Mailbox, Schema as DispatcherSchema},
+    dispatcher::{Dispatcher, Error as DispatcherError, Schema as DispatcherSchema},
     error::{
         CallSite, CallType, ErrorKind, ErrorMatch, ExecutionError, ExecutionFail, ExecutionStatus,
     },
@@ -151,6 +151,7 @@ use std::fmt;
 
 use exonum_merkledb::{BinaryValue, Fork, Snapshot};
 
+use self::dispatcher::Mailbox;
 use crate::{
     blockchain::Blockchain,
     crypto::{Hash, PublicKey},
